fix(05): split project pages into server/client for CSP nonce support
