+++ src/app/projects/partner-performance/page.tsx
@@ -1,140 +1,31 @@
-'use client'
+import { headers } from 'next/headers'
+import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
+import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
+import { siteConfig } from '@/lib/site'
+import { PartnerContent } from './_components/PartnerContent'
 
-import { TrendingUp, Users, Target, DollarSign } from 'lucide-react'
-import { useQueryState } from 'nuqs'
-
-import { safeJsonLdStringify } from '@/lib/json-ld-utils'
-import { ProjectPageLayout } from '@/components/projects/project-page-layout'
-import { MetricsGrid } from '@/components/projects/metrics-grid'
-import { SectionCard } from '@/components/ui/section-card'
-import { formatCurrency, formatNumber, formatPercentage } from '@/lib/data-formatters'
-import { NarrativeSections } from './_components/NarrativeSections'
-import { OverviewTab } from './_components/OverviewTab'
-import { TiersTab } from './_components/TiersTab'
-import { TopPerformersTab } from './_components/TopPerformersTab'
-import { StrategicInsights } from './_components/StrategicInsights'
-
-// Real data based on CSV analysis: 83.2% win rate, $1.1M partner revenue
-const partnerMetrics = {
-  totalPartners: 47,
-  activePartners: 34,
-  partnerRevenue: 904387,
-  partnerContribution: 83.2,
-  winRate: 83.2,
-  quickRatio: 4.7, // SaaS benchmark >4
-}
-
-const tabs = ['overview', 'tiers', 'top-performers'] as const
-type Tab = (typeof tabs)[number]
-
-export default function PartnerPerformanceIntelligence() {
-  const [activeTab, setActiveTab] = useQueryState('tab', { defaultValue: 'overview' as Tab })
-
-  // Standardized metrics configuration using consistent data formatting
-  const metrics = [
-    {
-      id: 'partner-revenue',
-      icon: DollarSign,
-      label: 'Partner Revenue',
-      value: formatCurrency(partnerMetrics.partnerRevenue, { compact: true }),
-      subtitle: `${formatPercentage(partnerMetrics.partnerContribution / 100)} of Total Revenue`,
-      variant: 'primary' as const,
-    },
-    {
-      id: 'win-rate',
-      icon: Target,
-      label: 'Win Rate',
-      value: formatPercentage(partnerMetrics.winRate / 100),
-      subtitle: 'Partner Channel Efficiency',
-      variant: 'secondary' as const,
-    },
-    {
-      id: 'active-partners',
-      icon: Users,
-      label: 'Active Partners',
-      value: formatNumber(partnerMetrics.activePartners),
-      subtitle: `of ${formatNumber(partnerMetrics.totalPartners)} Total`,
-      variant: 'primary' as const,
-    },
-    {
-      id: 'quick-ratio',
-      icon: TrendingUp,
-      label: 'Quick Ratio',
-      value: `${partnerMetrics.quickRatio}x`,
-      subtitle: 'SaaS Benchmark: >4.0x',
-      variant: 'secondary' as const,
-    },
-  ]
+export default async function PartnerPerformanceIntelligence() {
+  const nonce = (await headers()).get('x-nonce')
 
   return (
     <>
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: safeJsonLdStringify({
-          '@context': 'https://schema.org',
-          '@type': 'CreativeWork',
-          name: 'Partner Performance Analytics',
-          description: 'Strategic channel analytics and partner ROI intelligence demonstrating 83.2% win rate across multi-tier partner ecosystem with real-time performance tracking.',
-          url: 'https://richardwhudsonjr.com/projects/partner-performance',
-          author: { '@type': 'Person', name: 'Richard Hudson', url: 'https://richardwhudsonjr.com' },
-          genre: 'Partnership Program',
-          keywords: 'Partner Analytics, Channel Performance, Partnership Program, Revenue Operations',
-          inLanguage: 'en-US',
-          isAccessibleForFree: true,
-        }) }}
+      <ProjectJsonLd
+        title="Partner Performance Analytics"
+        description="Strategic channel analytics and partner ROI intelligence demonstrating 83.2% win rate across multi-tier partner ecosystem with real-time performance tracking."
+        slug="partner-performance"
+        category="Partnership Program"
+        tags={['Partner Analytics', 'Channel Performance', 'Partnership Program', 'Revenue Operations']}
+        nonce={nonce}
       />
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: safeJsonLdStringify({
-          '@context': 'https://schema.org',
-          '@type': 'BreadcrumbList',
-          itemListElement: [
-            { '@type': 'ListItem', position: 1, name: 'Home', item: 'https://richardwhudsonjr.com' },
-            { '@type': 'ListItem', position: 2, name: 'Projects', item: 'https://richardwhudsonjr.com/projects' },
-            { '@type': 'ListItem', position: 3, name: 'Partner Performance Analytics', item: 'https://richardwhudsonjr.com/projects/partner-performance' },
-          ],
-        }) }}
+      <BreadcrumbListJsonLd
+        items={[
+          { name: 'Home', url: siteConfig.url },
+          { name: 'Projects', url: `${siteConfig.url}/projects` },
+          { name: 'Partner Performance Analytics', url: `${siteConfig.url}/projects/partner-performance` },
+        ]}
+        nonce={nonce}
       />
-    <ProjectPageLayout
-      title="Partner Performance Intelligence Dashboard"
-      description="Strategic channel analytics and partner ROI intelligence demonstrating 83.2% win rate across multi-tier partner ecosystem. Real-time performance tracking following industry-standard 80/20 partner revenue distribution."
-      tags={[
-        { label: `Channel ROI: ${partnerMetrics.quickRatio}x`, variant: 'primary' },
-        {
-          label: `Win Rate: ${formatPercentage(partnerMetrics.winRate / 100)}`,
-          variant: 'secondary',
-        },
-        { label: 'Partner Intelligence', variant: 'primary' },
-        { label: 'Revenue Operations', variant: 'secondary' },
-      ]}
-      showTimeframes={true}
-      timeframes={tabs.map((t) => t.charAt(0).toUpperCase() + t.slice(1).replace('-', ' '))}
-      activeTimeframe={activeTab.charAt(0).toUpperCase() + activeTab.slice(1).replace('-', ' ')}
-      onTimeframeChange={(timeframe) => {
-        const tab = timeframe.toLowerCase().replace(' ', '-') as Tab
-        setActiveTab(tab)
-      }}
-    >
-      {/* Key Metrics using standardized MetricsGrid */}
-      <MetricsGrid metrics={metrics} columns={4} className="mb-8" />
-
-      {/* Tab Content wrapped in SectionCard */}
-      <SectionCard
-        title="Partner Performance Analysis"
-        description="Comprehensive partner analytics and performance insights"
-        className="mb-8"
-      >
-        {activeTab === 'overview' && <OverviewTab />}
-        {activeTab === 'tiers' && <TiersTab />}
-        {activeTab === 'top-performers' && <TopPerformersTab />}
-      </SectionCard>
-
-      {/* Strategic Insights */}
-      <StrategicInsights winRate={partnerMetrics.winRate} quickRatio={partnerMetrics.quickRatio} />
-
-      {/* Professional Narrative Sections - STAR Method */}
-      <NarrativeSections />
-    </ProjectPageLayout>
+      <PartnerContent />
     </>
   )
 }
