+++ src/app/projects/revenue-operations-center/page.tsx
@@ -1,160 +1,31 @@
-'use client'
+import { headers } from 'next/headers'
+import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
+import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
+import { siteConfig } from '@/lib/site'
+import { RevOpsCenterContent } from './_components/RevOpsCenterContent'
 
-import { Suspense, useMemo } from 'react'
-import nextDynamic from 'next/dynamic'
-import { DollarSign, Target, BarChart3, Users, Activity } from 'lucide-react'
-import { useQueryState } from 'nuqs'
-
-import { safeJsonLdStringify } from '@/lib/json-ld-utils'
-import { ProjectPageLayout } from '@/components/projects/project-page-layout'
-import { revenueMetrics } from './data/constants'
-import { formatCurrency, formatPercent } from './utils'
-import { MetricsGrid } from '@/components/projects/metrics-grid'
-import { KPIAlerts } from './_components/KPIAlerts'
-import { StrategicImpact } from './_components/StrategicImpact'
-import { NarrativeSections } from './_components/NarrativeSections'
-
-// Tab loading skeleton
-function TabSkeleton() {
-  return (
-    <div className="glass rounded-2xl p-8 mb-12">
-      <div className="h-8 w-64 bg-muted animate-pulse rounded mb-4" />
-      <div className="h-[300px] w-full bg-muted animate-pulse rounded-lg" />
-    </div>
-  )
-}
-
-// Lazy load tab components - only active tab loads
-const OverviewTab = nextDynamic(() => import('./_components/OverviewTab').then(m => ({ default: m.OverviewTab })), {
-  loading: () => <TabSkeleton />,
-  ssr: true,
-})
-const PipelineTab = nextDynamic(() => import('./_components/PipelineTab').then(m => ({ default: m.PipelineTab })), {
-  loading: () => <TabSkeleton />,
-  ssr: true,
-})
-const ForecastingTab = nextDynamic(() => import('./_components/ForecastingTab').then(m => ({ default: m.ForecastingTab })), {
-  loading: () => <TabSkeleton />,
-  ssr: true,
-})
-const OperationsTab = nextDynamic(() => import('./_components/OperationsTab').then(m => ({ default: m.OperationsTab })), {
-  loading: () => <TabSkeleton />,
-  ssr: true,
-})
-
-const tabs = ['overview', 'pipeline', 'forecasting', 'operations'] as const
-type Tab = (typeof tabs)[number]
-
-export default function RevenueOperationsCenter() {
-  const [activeTab, setActiveTab] = useQueryState('tab', { defaultValue: 'overview' as Tab })
-
-  // Memoize metrics to prevent recreation on every render
-  const metrics = useMemo(() => [
-    {
-      id: 'total-revenue',
-      icon: DollarSign,
-      label: 'Total Revenue',
-      value: formatCurrency(revenueMetrics.totalRevenue),
-      subtitle: `+${formatPercent(revenueMetrics.revenueGrowth)} YoY`,
-      variant: 'primary' as const,
-    },
-    {
-      id: 'forecast-accuracy',
-      icon: Target,
-      label: 'Forecast',
-      value: formatPercent(revenueMetrics.forecastAccuracy),
-      subtitle: 'Accuracy Rate',
-      variant: 'primary' as const,
-    },
-    {
-      id: 'pipeline-health',
-      icon: BarChart3,
-      label: 'Pipeline',
-      value: formatPercent(revenueMetrics.pipelineHealth),
-      subtitle: 'Health Score',
-      variant: 'secondary' as const,
-    },
-    {
-      id: 'active-deals',
-      icon: Users,
-      label: 'Active Deals',
-      value: revenueMetrics.activeDeals.toString(),
-      subtitle: `${formatCurrency(revenueMetrics.avgDealSize)} avg`,
-      variant: 'primary' as const,
-    },
-    {
-      id: 'target-attainment',
-      icon: Activity,
-      label: 'Target',
-      value: formatPercent(revenueMetrics.targetAttainment),
-      subtitle: 'Attainment',
-      variant: 'primary' as const,
-    },
-  ], [])
+export default async function RevenueOperationsCenter() {
+  const nonce = (await headers()).get('x-nonce')
 
   return (
     <>
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: safeJsonLdStringify({
-          '@context': 'https://schema.org',
-          '@type': 'CreativeWork',
-          name: 'Revenue Operations Center',
-          description: 'Comprehensive revenue operations dashboard consolidating pipeline health, forecasting accuracy, partner performance, and operational KPIs with 96.8% forecast accuracy.',
-          url: 'https://richardwhudsonjr.com/projects/revenue-operations-center',
-          author: { '@type': 'Person', name: 'Richard Hudson', url: 'https://richardwhudsonjr.com' },
-          genre: 'Revenue Operations',
-          keywords: 'Revenue Operations, RevOps Dashboard, Sales Analytics, Business Intelligence',
-          inLanguage: 'en-US',
-          isAccessibleForFree: true,
-        }) }}
+      <ProjectJsonLd
+        title="Revenue Operations Center"
+        description="Comprehensive revenue operations dashboard consolidating pipeline health, forecasting accuracy, partner performance, and operational KPIs with 96.8% forecast accuracy."
+        slug="revenue-operations-center"
+        category="Revenue Operations"
+        tags={['Revenue Operations', 'RevOps Dashboard', 'Sales Analytics', 'Business Intelligence']}
+        nonce={nonce}
       />
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: safeJsonLdStringify({
-          '@context': 'https://schema.org',
-          '@type': 'BreadcrumbList',
-          itemListElement: [
-            { '@type': 'ListItem', position: 1, name: 'Home', item: 'https://richardwhudsonjr.com' },
-            { '@type': 'ListItem', position: 2, name: 'Projects', item: 'https://richardwhudsonjr.com/projects' },
-            { '@type': 'ListItem', position: 3, name: 'Revenue Operations Center', item: 'https://richardwhudsonjr.com/projects/revenue-operations-center' },
-          ],
-        }) }}
+      <BreadcrumbListJsonLd
+        items={[
+          { name: 'Home', url: siteConfig.url },
+          { name: 'Projects', url: `${siteConfig.url}/projects` },
+          { name: 'Revenue Operations Center', url: `${siteConfig.url}/projects/revenue-operations-center` },
+        ]}
+        nonce={nonce}
       />
-    <ProjectPageLayout
-      title="Revenue Operations Command Center"
-      description="Comprehensive revenue operations dashboard consolidating pipeline health, forecasting accuracy, partner performance, and operational KPIs. Real-time insights with 96.8% forecast accuracy and 89.7% operational efficiency across sales, marketing, and partner channels."
-      tags={[
-        { label: 'Forecast Accuracy: 96.8%', variant: 'primary' },
-        { label: 'Pipeline Health: 92.4%', variant: 'secondary' },
-        { label: 'Revenue Growth: +34.2%', variant: 'primary' },
-        { label: 'Operations Dashboard', variant: 'secondary' },
-      ]}
-      showTimeframes={true}
-      timeframes={tabs.map((t) => t.charAt(0).toUpperCase() + t.slice(1))}
-      activeTimeframe={activeTab.charAt(0).toUpperCase() + activeTab.slice(1)}
-      onTimeframeChange={(timeframe) => setActiveTab(timeframe.toLowerCase() as Tab)}
-    >
-          {/* Key Metrics using standardized MetricsGrid */}
-          <MetricsGrid metrics={metrics} columns={4} className="mb-12" />
-
-          {/* KPI Alerts */}
-          <KPIAlerts />
-
-          {/* Tab Content with Suspense */}
-          <Suspense fallback={<TabSkeleton />}>
-            {activeTab === 'overview' && <OverviewTab />}
-            {activeTab === 'pipeline' && <PipelineTab />}
-            {activeTab === 'forecasting' && <ForecastingTab />}
-            {activeTab === 'operations' && <OperationsTab />}
-          </Suspense>
-
-          {/* Strategic Impact */}
-          <StrategicImpact />
-
-          {/* Professional Narrative Sections */}
-          <NarrativeSections />
-    </ProjectPageLayout>
+      <RevOpsCenterContent />
     </>
   )
 }
